spark-server/requirements.txt
fastapi
pyspark
python-jose[cryptography]
passlib
